llamaindexer/code/requirement.txt





# LlamaIndex
llama-index-core>=0.12.8
llama-index-llms-ibm>=0.3.0
llama-index-embeddings-ibm>=0.3.0
llama-index-readers-file>=0.4.1
llama-index-vector-stores-chroma>=0.4.1
llama-index-embeddings-huggingface
llama-index-llms-ollama

# Chromadb
chromadb==0.5.23

# ファイル読み込み用
python-docx>=1.1.2
docx2txt>=0.8
transformers<4.50
python-pptx
Pillow

# others
python-dotenv>=1.0.1
pandas>=2.1.4
nltk